Support reading really big comic archives when building metadata
Previously big files (>~100mb) would cause the process to run out of memory.

This is because the whole file was read, only to extract the cover image.

Now that @comix/parser can read by file name directly (by streaming only what
is needed) this isn't necessary, and we can just give the parser a file path.